Completed todos and bugfixes  also migrated from Ivy on Gradle
