+++ CSSInject.ipr
@@ -1,0 +1,171 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ArtifactManager">
+    <artifact type="war" name="Gradle : org.vaadin:addon : addon-2.0.4.war">
+      <output-path>$PROJECT_DIR$/build/classes/artifacts/addon</output-path>
+      <properties id="gradle-properties">
+        <options external-project-path="$PROJECT_DIR$/addon" />
+      </properties>
+      <root id="archive" name="addon-2.0.4.war">
+        <element id="javaee-facet-resources" facet="addon/web/Web Gradle : addon-2.0.4.war" />
+        <element id="directory" name="WEB-INF">
+          <element id="directory" name="classes">
+            <element id="module-output" name="addon" />
+            <element id="dir-copy" path="$PROJECT_DIR$/addon/build/resources/main" />
+          </element>
+        </element>
+        <element id="directory" name="META-INF">
+          <element id="file-copy" path="$PROJECT_DIR$/build/classes/artifacts/addon/exploded/addon-2.0.4.war/META-INF/MANIFEST.MF" />
+        </element>
+      </root>
+    </artifact>
+    <artifact type="exploded-war" name="Gradle : org.vaadin:addon : addon-2.0.4.war (exploded)">
+      <output-path>$PROJECT_DIR$/build/classes/artifacts/addon/exploded/addon-2.0.4.war</output-path>
+      <properties id="gradle-properties">
+        <options external-project-path="$PROJECT_DIR$/addon" />
+      </properties>
+      <root id="root">
+        <element id="javaee-facet-resources" facet="addon/web/Web Gradle : addon-2.0.4.war" />
+        <element id="directory" name="WEB-INF">
+          <element id="directory" name="classes">
+            <element id="module-output" name="addon" />
+            <element id="dir-copy" path="$PROJECT_DIR$/addon/build/resources/main" />
+          </element>
+        </element>
+        <element id="directory" name="META-INF">
+          <element id="file-copy" path="$PROJECT_DIR$/build/classes/artifacts/addon/exploded/addon-2.0.4.war/META-INF/MANIFEST.MF" />
+        </element>
+      </root>
+    </artifact>
+    <artifact type="war" name="Gradle : org.vaadin:demo : CssInject.war">
+      <output-path>$PROJECT_DIR$/build/classes/artifacts/demo</output-path>
+      <properties id="gradle-properties">
+        <options external-project-path="$PROJECT_DIR$/demo" />
+      </properties>
+      <root id="archive" name="CssInject.war">
+        <element id="javaee-facet-resources" facet="demo/web/Web Gradle : CssInject.war" />
+        <element id="directory" name="WEB-INF">
+          <element id="directory" name="classes">
+            <element id="module-output" name="demo" />
+            <element id="dir-copy" path="$PROJECT_DIR$/demo/build/resources/main" />
+          </element>
+          <element id="directory" name="lib">
+            <element id="archive" name="org.vaadin-addon-2.0.4.jar">
+              <element id="module-output" name="addon" />
+              <element id="dir-copy" path="$PROJECT_DIR$/addon/build/resources/main" />
+            </element>
+          </element>
+        </element>
+        <element id="directory" name="META-INF">
+          <element id="file-copy" path="$PROJECT_DIR$/build/classes/artifacts/demo/exploded/CSSInject.war/META-INF/MANIFEST.MF" />
+        </element>
+      </root>
+    </artifact>
+    <artifact type="exploded-war" name="Gradle : org.vaadin:demo : CssInject.war (exploded)">
+      <output-path>$PROJECT_DIR$/build/classes/artifacts/demo/exploded/CssInject.war</output-path>
+      <properties id="gradle-properties">
+        <options external-project-path="$PROJECT_DIR$/demo" />
+      </properties>
+      <root id="root">
+        <element id="javaee-facet-resources" facet="demo/web/Web Gradle : CssInject.war" />
+        <element id="directory" name="WEB-INF">
+          <element id="directory" name="classes">
+            <element id="module-output" name="demo" />
+            <element id="dir-copy" path="$PROJECT_DIR$/demo/build/resources/main" />
+          </element>
+          <element id="directory" name="lib">
+            <element id="archive" name="org.vaadin-addon-2.0.4.jar">
+              <element id="module-output" name="addon" />
+              <element id="dir-copy" path="$PROJECT_DIR$/addon/build/resources/main" />
+            </element>
+          </element>
+        </element>
+        <element id="directory" name="META-INF">
+          <element id="file-copy" path="$PROJECT_DIR$/build/classes/artifacts/demo/exploded/CSSInject.war/META-INF/MANIFEST.MF" />
+        </element>
+      </root>
+    </artifact>
+  </component>
+  <component name="CompilerConfiguration">
+    <resourceExtensions />
+    <wildcardResourcePatterns>
+      <entry name="!?*.java" />
+      <entry name="!?*.form" />
+      <entry name="!?*.class" />
+      <entry name="!?*.groovy" />
+      <entry name="!?*.scala" />
+      <entry name="!?*.flex" />
+      <entry name="!?*.kt" />
+      <entry name="!?*.clj" />
+      <entry name="!?*.aj" />
+    </wildcardResourcePatterns>
+    <annotationProcessing>
+      <profile default="true" name="Default" enabled="false">
+        <processorPath useClasspath="true" />
+      </profile>
+    </annotationProcessing>
+  </component>
+  <component name="CopyrightManager" default="" />
+  <component name="EntryPointsManager">
+    <entry_points version="2.0" />
+  </component>
+  <component name="FrameworkDetectionExcludesConfiguration">
+    <file type="web" url="file://$PROJECT_DIR$" />
+  </component>
+  <component name="GradleSettings">
+    <option name="linkedExternalProjectsSettings">
+      <GradleProjectSettings>
+        <option name="distributionType" value="DEFAULT_WRAPPED" />
+        <option name="externalProjectPath" value="$PROJECT_DIR$" />
+        <option name="gradleJvm" value="#JAVA_HOME" />
+        <option name="modules">
+          <set>
+            <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/addon" />
+            <option value="$PROJECT_DIR$/demo" />
+          </set>
+        </option>
+        <option name="useAutoImport" value="true" />
+      </GradleProjectSettings>
+    </option>
+  </component>
+  <component name="ProjectLevelVcsManager" settingsEditedManually="false">
+    <OptionsSetting value="true" id="Add" />
+    <OptionsSetting value="true" id="Remove" />
+    <OptionsSetting value="true" id="Checkout" />
+    <OptionsSetting value="true" id="Update" />
+    <OptionsSetting value="true" id="Status" />
+    <OptionsSetting value="true" id="Edit" />
+    <ConfirmationsSetting value="0" id="Add" />
+    <ConfirmationsSetting value="0" id="Remove" />
+  </component>
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/CSSInject.iml" filepath="$PROJECT_DIR$/CSSInject.iml" />
+      <module fileurl="file://$PROJECT_DIR$/addon/addon.iml" filepath="$PROJECT_DIR$/addon/addon.iml" />
+      <module fileurl="file://$PROJECT_DIR$/demo/demo.iml" filepath="$PROJECT_DIR$/demo/demo.iml" />
+    </modules>
+  </component>
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_6" default="true" assert-keyword="true" jdk-15="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
+    <output url="file://$PROJECT_DIR$/build/classes" />
+  </component>
+  <component name="VcsDirectoryMappings">
+    <mapping directory="" vcs="Git" />
+  </component>
+  <component name="masterDetails">
+    <states>
+      <state key="ProjectJDKs.UI">
+        <settings>
+          <last-edited>1.7</last-edited>
+          <splitter-proportions>
+            <option name="proportions">
+              <list>
+                <option value="0.2" />
+              </list>
+            </option>
+          </splitter-proportions>
+        </settings>
+      </state>
+    </states>
+  </component>
+</project>
