--- WebContent/META-INF/MANIFEST.MF
@@ -1,7 +1,0 @@
-Manifest-Version: 1.0
-Implementation-Title: CSSInject
-Implementation-Version: 2.0.3
-Vaadin-Package-Version: 1
-Vaadin-Widgetsets: org.vaadin.cssinject.Cssinject_addonWidgetset
-Class-Path: 
-
