--- WebContent/WEB-INF/web.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app version="2.4" xmlns="http://java.sun.com/xml/ns/j2ee"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/web-app_2_4.xsd"
-	id="WebApp_ID">
-	<display-name>CSSInject</display-name>
-	<context-param>
-		<description>
-		Vaadin production mode</description>
-		<param-name>productionMode</param-name>
-		<param-value>false</param-value>
-	</context-param>
-	<servlet>
-		<servlet-name>CSSInject</servlet-name>
-		<servlet-class>com.vaadin.server.VaadinServlet</servlet-class>
-		<init-param>
-			<description>
-			Vaadin UI class to start</description>
-			<param-name>UI</param-name>
-			<param-value>org.vaadin.cssinject.demo.CSSInjectDemo</param-value>
-		</init-param>
-		<init-param>
-			<description>
-			Application widgetset</description>
-			<param-name>widgetset</param-name>
-			<param-value>org.vaadin.cssinject.Cssinject_addonWidgetset</param-value>
-		</init-param>
-	</servlet>
-	<servlet-mapping>
-		<servlet-name>CSSInject</servlet-name>
-		<url-pattern>/*</url-pattern>
-	</servlet-mapping>
-	<welcome-file-list>
-		<welcome-file>index.html</welcome-file>
-		<welcome-file>index.htm</welcome-file>
-		<welcome-file>index.jsp</welcome-file>
-		<welcome-file>default.html</welcome-file>
-		<welcome-file>default.htm</welcome-file>
-		<welcome-file>default.jsp</welcome-file>
-	</welcome-file-list>
-</web-app>
