+++ addon/src/main/resources/org/vaadin/cssinject/CssInjectWidgetSet.gwt.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE module PUBLIC "-//Vaadin//DTD Vaadin 7//EN" "https://raw.github.com/vaadin/gwt/master/distro-source/core/src/gwt-module.dtd">
+<!-- WS Compiler: manually edited -->
+<module>
+        <inherits name="com.vaadin.DefaultWidgetSet" />
+        <set-property name="user.agent" value="ie8,ie9,gecko1_8,safari,ie10" />
+        <source path="client" />
+        <source path="shared" />
+        <collapse-all-properties />
+        <set-property name="compiler.useSymbolMaps" value="true" />
+</module>
