--- src/org/vaadin/cssinject/Cssinject_addonWidgetset.gwt.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 1.7.0//EN" "http://google-web-toolkit.googlecode.com/svn/tags/1.7.0/distro-source/core/src/gwt-module.dtd">
-<module>
-	<inherits name="com.vaadin.DefaultWidgetSet" />
-	
-	<source path="client"></source>
-	<source path="shared"></source>
-	
-    <!--
-     Uncomment the following to compile the widgetset for one browser only.
-     This can reduce the GWT compilation time significantly when debugging.
-     The line should be commented out before deployment to production
-     environments.
-      
-     Multiple browsers can be specified for GWT 1.7 as a comma separated
-     list. The supported user agents at the moment of writing were:
-     ie6,ie8,gecko,gecko1_8,safari,opera
-     
-     The value gecko1_8 is used for Firefox 3 and later and safari is used for
-     webkit based browsers including Google Chrome.
-    -->
-<!--     <set-property name="user.agent" value="safari"/> -->
-
-</module>
