--- src/org/vaadin/cssinject/shared/CSSInjectState.java
@@ -1,13 +1,0 @@
-package org.vaadin.cssinject.shared;
-
-import com.vaadin.shared.communication.SharedState;
-
-public class CSSInjectState extends SharedState {
-
-    private static final long serialVersionUID = 8236742409269917338L;
-
-    public String styles;
-
-    public int styleSheetId = -1;
-
-}
